feat: To make it usable in SSR apps refactor platform and user agent constants and keyHandlers
Reorganized the constants in source/Constants.ts and keyHandlers in source/KeyHandlers.ts to improve readability and maintainability. Introduced a constants object and a function (getConstants) to initialize and retrieve the constants dynamically. Updated source/Clipboard.ts, source/Space.ts, source/MergeSplit.ts, source/KeyHandler.ts files accordingly to reflect these changes.

BREAKING CHANGE: Users of the constants should now import the getConstants function and use it to retrieve the constants and the KeyHandlers should now import the getKeyHandlers function and use it to retrieve the KeyHandlers.